Move resolveApiUrl from time.ts to auth.ts
resolveApiUrl follows the same flag → env → config → default pattern as
resolveApiKey, so it belongs alongside it in auth.ts. Added tests for
resolveApiUrl in the auth test suite.

Closes #32
